Refactor Meteor.isServer in server folder at root
