+++ lib/collections.js
@@ -66,13 +66,13 @@ Schema.AddressSchema =new SimpleSchema({
   },
   state: {
     type: String,
-    // Source d'erreur pour le package. 
+    // Source d'erreur pour le package.
   //  regEx: /^A[LKSZRAEP]|C[AOT]|D[EC]|F[LM]|G[AU]|HI|I[ADLN]|K[SY]|LA|M[ADEHINOPST]|N[CDEHJMVY]|O[HKR]|P[ARW]|RI|S[CD]|T[NX]|UT|V[AIT]|W[AIVY]$/,
     optional: true
   },
   zip: {
     type: String,
-        // Source d'erreur pour le package. 
+        // Source d'erreur pour le package.
     //regEx: /^[0-9]{5}$/,
     optional: true
   },
@@ -84,7 +84,7 @@ Schema.AddressSchema =new SimpleSchema({
 
 
 Schema.UserProfile = new SimpleSchema({
-    
+
    firstName: {
     type: String,
     label: 'First Name',
@@ -183,7 +183,7 @@ Schema.User = new SimpleSchema({
         blackbox: true
     }
 
-  }); 
+  });
 
 
 
@@ -216,7 +216,7 @@ Schema.MAILBOX = new SimpleSchema({
     status: {
     type: String
     }
-  }); 
+  });
 
 Schema.CHAT = new SimpleSchema({
     mailbox_Id: {
@@ -241,63 +241,8 @@ Schema.CHAT = new SimpleSchema({
       'label-type': 'floating'
     }
   },
-}); 
+});
 
 if (Meteor.isClient) {
   Meteor.subscribe('mailbox');
 }
-
-if (Meteor.isServer) {
-  //autorise l'envoi de mails...
-process.env.MAIL_URL="smtp://thecolibry%40gmail.com:C@mp1267.*@smtp.gmail.com:465/";
-// avant 465
-
-Meteor.methods({
-  sendEmail: function (to, from, subject, text) {
-    check([to, from, subject, text], [String]);
-
-    // Let other method calls from the same client start running,
-    // without waiting for the email sending to complete.
-    this.unblock();
-
-    Email.send({
-      to: to,
-      from: from,
-      subject: subject,
-      text: text
-    });
-
-    
-  }
-});
-
-
-
-
-Meteor.publish('images',function(){
-    var currentUserId = this.userId;
-    return IMAGES.find();
-  });
-
-Meteor.publish('mailbox',function(){
-    var currentUserId = this.userId;
-    return MAILBOX.find({ $or: [{ sender_Id: currentUserId }, {recipient_Id: currentUserId } ] });
-  });
-
-
-IMAGES.allow({
-  'insert': function (userId, doc) {
-    // add custom authentication code here
-    return true;
-  },
-  update: function(userId, doc, fieldNames, modifier) {
-  return true;
-  },
-  'download': function (userId) {
-    // add custom authentication code here
-    return true;
-}
-});
-}
-
-
